# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Assessment Node"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Retrieve the auth token",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return a valid Bearer access token for the valid client_credentials provided. The token has a time to live equal to expires_in",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "token",
                    "type",
                    "expires_in"
                  ],
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/policies": {
      "summary": "Get the list of policies client",
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "Get the list of policies client paginated and limited to 10 elements by default.",
        "description": "Can be accessed by client with role user (it will retrieve its own policies) and admin (it will retrieve all the policies)",
        "operationId": "getPolicies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "properties": {
                "limit": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "properties": {
                "limit": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of policies client",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "amountInsured": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "inceptionDate": {
                        "type": "string"
                      },
                      "installmentPayment": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/policies/{id}": {
      "summary": "Get the details of a policys client",
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "Get the details of a policys client",
        "description": "Can be accessed by client with role user (it will retrieve its own policy) and admin (it will retrieve all the policies)",
        "operationId": "getPolicyById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get the details of a policys client",
            "content": {
              "application/json": {
                "schema": {
                  "id": {
                    "type": "string"
                  },
                  "amountInsured": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "inceptionDate": {
                    "type": "string"
                  },
                  "installmentPayment": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clients": {
      "summary": "Get the list of clients details",
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "Get the list of clients details paginated and limited to 10 elements by default. This API endpoint access also an optional filter query to filter by client name.",
        "description": "Can be accessed by client with role user (it will retrieve its own client details as only element of the list) and admin (it will retrieve all the clients list)",
        "operationId": "getClients",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "properties": {
                "limit": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "properties": {
                "limit": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "properties": {
                "limit": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of clients details",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      },
                      "policies": {
                        "items": {
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "amountInsured": {
                              "type": "string"
                            },
                            "inceptionDate": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{id}": {
      "summary": "Get the clients details",
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "Get the clients details",
        "description": "Can be accessed by client with role user (it will retrieve its own client details) and admin (it will retrieve any client details)",
        "operationId": "getClientById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clients details",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      },
                      "policies": {
                        "items": {
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "amountInsured": {
                              "type": "string"
                            },
                            "inceptionDate": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{id}/policies": {
      "summary": "Get the clients policies",
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "Get the clients policies",
        "description": "Can be accessed by client with role user (it will retrieve its own client policy list) and admin (it will retrieve any client policy list)",
        "operationId": "getClientPoliciesById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            },
            "required": true
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "properties": {
                "limit": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "properties": {
                "limit": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clients policies",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "amountInsured": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "inceptionDate": {
                        "type": "string"
                      },
                      "installmentPayment": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}